+++ MetaQuest/Editor/MetaQuestFeatureEditor.cs
@@ -10,6 +10,14 @@ namespace UnityEditor.XR.OpenXR.Features.MetaQuestSupport
     [CustomEditor(typeof(MetaQuestFeature))]
     internal class MetaQuestFeatureEditor : Editor
     {
+        private bool m_ShowAndroidExperimental = false;
+        private bool m_LateLatchingModeEnabled;
+        private bool m_LateLatchingDebug;
+
+        private static GUIContent s_LateLatchingSupportedLabel = EditorGUIUtility.TrTextContent("Late Latching (Vulkan)");
+        private static GUIContent s_LateLatchingDebugLabel = EditorGUIUtility.TrTextContent("Late Latching Debug Mode");
+        private static GUIContent s_ShowAndroidExperimentalLabel = EditorGUIUtility.TrTextContent("Experimental", "Experimental settings that are under active development and should be used with caution.");
+
         struct TargetDeviceProperty
         {
             public SerializedProperty property;
@@ -19,6 +27,12 @@ namespace UnityEditor.XR.OpenXR.Features.MetaQuestSupport
         private List<TargetDeviceProperty> targetDeviceProperties;
         private Dictionary<string, bool> activeTargetDevices;
         private SerializedProperty forceRemoveInternetPermission;
+        private SerializedProperty systemSplashScreen;
+
+        private SerializedProperty symmetricProjection;
+
+        private SerializedProperty m_LateLatchingModeProperty;
+        private SerializedProperty m_LateLatchingDebugProperty;
 
         void InitActiveTargetDevices()
         {
@@ -40,12 +54,21 @@ namespace UnityEditor.XR.OpenXR.Features.MetaQuestSupport
         {
             forceRemoveInternetPermission =
                 serializedObject.FindProperty("forceRemoveInternetPermission");
+            systemSplashScreen =
+                serializedObject.FindProperty("systemSplashScreen");
+
+            symmetricProjection =
+                serializedObject.FindProperty("symmetricProjection");
 
             targetDeviceProperties = new List<TargetDeviceProperty>();
             InitActiveTargetDevices();
             if (activeTargetDevices.Count == 0)
                 return;
             var targetDevicesProperty = serializedObject.FindProperty("targetDevices");
+
+            // mapping these to Properties so tha we can get Undo/redo functionality
+            m_LateLatchingDebugProperty = serializedObject.FindProperty("lateLatchingDebug");
+            m_LateLatchingModeProperty = serializedObject.FindProperty("lateLatchingMode");
 
             for (int i = 0; i < targetDevicesProperty.arraySize; ++i)
             {
@@ -64,18 +87,52 @@ namespace UnityEditor.XR.OpenXR.Features.MetaQuestSupport
 
         public override void OnInspectorGUI()
         {
+            // Update anything from the serializable object
+            EditorGUIUtility.labelWidth = 215.0f;
+
             serializedObject.Update();
+            EditorGUILayout.LabelField("Rendering Settings", EditorStyles.boldLabel);
+            EditorGUILayout.PropertyField(symmetricProjection, new GUIContent("Symmetric Projection (Vulkan)"));
+
+            EditorGUILayout.Space();
+
             EditorGUILayout.LabelField("Manifest Settings", EditorStyles.boldLabel);
             EditorGUILayout.PropertyField(forceRemoveInternetPermission);
+            EditorGUILayout.PropertyField(systemSplashScreen);
+
+            EditorGUILayout.Space();
 
             EditorGUILayout.LabelField("Target Devices", EditorStyles.boldLabel);
 
+            // Layout the Target Device properties
+            EditorGUI.indentLevel++;
             foreach (var device in targetDeviceProperties)
             {
                 EditorGUILayout.PropertyField(device.property, device.label);
             }
+            EditorGUI.indentLevel--;
 
+            // Foldout for the Experimental properties
+            if (m_ShowAndroidExperimental = EditorGUILayout.Foldout(m_ShowAndroidExperimental, s_ShowAndroidExperimentalLabel, EditorStyles.miniBoldFont))
+            {
+                EditorGUI.indentLevel++;
+                EditorGUILayout.PropertyField(m_LateLatchingModeProperty, s_LateLatchingSupportedLabel);
+                EditorGUILayout.PropertyField(m_LateLatchingDebugProperty, s_LateLatchingDebugLabel);
+                EditorGUI.indentLevel--;
+            }
+
+            EditorGUIUtility.labelWidth = 0.0f;
+
+            // update any serializable properties
             serializedObject.ApplyModifiedProperties();
+
+            OpenXRSettings androidOpenXRSettings = OpenXRSettings.GetSettingsForBuildTargetGroup(BuildTargetGroup.Android);
+            var serializedOpenXrSettings = new SerializedObject(androidOpenXRSettings);
+
+            androidOpenXRSettings.symmetricProjection = symmetricProjection.boolValue;
+            serializedOpenXrSettings.ApplyModifiedProperties();
+
+            EditorGUIUtility.labelWidth = 0.0f;
         }
     }
 }