+++ MockRuntime/Native~/mock_runtime/mock_runtime.cpp
@@ -1373,6 +1373,17 @@ XrResult MockRuntime::CreateReferenceSpace(const XrReferenceSpaceCreateInfo* cre
     case XR_REFERENCE_SPACE_TYPE_UNBOUNDED_MSFT:
         break;
 
+    case XR_REFERENCE_SPACE_TYPE_LOCAL_FLOOR_EXT:
+        if ((createFlags & MR_CREATE_LOCAL_FLOOR_REFERENCE_SPACE_EXT) != 0)
+        {
+            break;
+        }
+        else
+        {
+            // Extension not enabled
+            return XR_ERROR_VALIDATION_FAILURE;
+        }
+
     default:
         return XR_ERROR_REFERENCE_SPACE_UNSUPPORTED;
     }